Here is a very useful metadata query
I used this query a lot to try to understand the data I was looking at. I still think variations of it can be helpful - especially when you don't have access to any third-party tools. Besides, it's just good practice if you're a beginner and good knowledge to have if you ever want to work in some metadata as part of a stored procedure you're writing.