execute this in a cursor for some real fun :)
The main purpose of this stored procedure is to execute this in another stored procedure containing a cursor and some temp tables, which would return all of the views in the current database and the number of other views that they reference (excluding columns). When referenced_minor_name is null, it represents the record for the parent entity (which, in our case, is the view - but could also be a table). referenced_class = 1 since views only belong to this class and we don't care about other classes, which are for referencing things like partitions