cursor that runs CountObjectReferences
It creates temp tables to store the results of the cursor that counts the number of views that are referenced for each view's definition.  For example, if you create a view that is joined to two other views in its definition one time each, then the "RefCount" for that view would be 2. The "Nesting" part is still a work in progress...